requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.1.1
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=h36ffca9_3
absl-py=2.1.0=pyhd8ed1ab_1
aiohttp=3.7.4.post0=py310he2412df_1
asgiref=3.8.1=py310haa95532_0
astunparse=1.6.3=pyhd8ed1ab_3
async-timeout=3.0.1=py_1000
attrs=25.1.0=pyh71513ae_0
blinker=1.9.0=pyhff2d567_0
brotli-python=1.0.9=py310h00ffb61_8
bzip2=1.0.8=h2bbff1b_6
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2025.2.25=haa95532_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.5.2=pyhd8ed1ab_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.15.1=py310h628cb3f_3
chardet=4.0.0=py310h5588dad_3
charset-normalizer=3.4.1=pyhd8ed1ab_0
click=8.1.8=pyh7428d3b_0
colorama=0.4.6=pyhd8ed1ab_1
cryptography=39.0.0=py310h52f42fa_0
django=5.1.3=py310haa95532_0
djangorestframework=3.15.2=pyhd8ed1ab_1
eigen=3.4.0=h2d74725_0
ffmpeg=4.2.3=ha925a31_0
flatbuffers=23.3.3=h63175ca_0
freetype=2.10.4=h546665d_1
gast=0.4.0=pyh9f0ad1d_0
gettext=0.21.1=h5728263_0
giflib=5.2.1=h64bf75a_3
glib=2.78.4=hd77b12b_0
glib-tools=2.78.4=hd77b12b_0
google-auth=2.38.0=pyhd8ed1ab_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyhd8ed1ab_2
grpc-cpp=1.48.2=hf108199_0
grpcio=1.48.2=py310hf108199_0
gst-plugins-base=1.18.5=he07aa86_0
gstreamer=1.18.5=hdff456e_3
h2=4.2.0=pyhd8ed1ab_0
h5py=3.7.0=nompi_py310h00cbb18_100
hdf5=1.12.1=nompi_h2a0e4a3_100
hpack=4.1.0=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=58.2=ha925a31_3
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.6.1=pyha770c72_0
intel-openmp=2024.2.1=h57928b3_1083
jbig=2.1=h8d14728_2003
jpeg=9e=hcfcfb64_3
keras=2.10.0=pyhd8ed1ab_0
krb5=1.19.4=h5b6d351_0
lerc=2.2.1=h0e60522_0
libabseil-static=20211102.0=cxx11_h58a5ce6_3
libblas=3.9.0=31_h641d27c_mkl
libcblas=3.9.0=31_h5e41251_mkl
libclang=14.0.6=default_hb5a9fac_2
libclang13=14.0.6=default_h8e68704_2
libcurl=8.9.1=h0416ee5_0
libdeflate=1.7=h8ffe710_5
libffi=3.4.4=hd77b12b_1
libglib=2.78.4=ha17d25a_0
libhwloc=2.9.1=h51c2c0f_0
libiconv=1.17=h8ffe710_0
liblapack=3.9.0=31_h1aa476e_mkl
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libssh2=1.10.0=h680486a_2
libtiff=4.3.0=h0c97f57_1
libvorbis=1.3.7=h0e60522_0
libwebp=1.3.0=hcfcfb64_0
libwebp-base=1.3.0=hcfcfb64_0
libxml2=2.10.4=h0ad7f3c_2
libxslt=1.1.37=h0192164_0
lz4-c=1.9.4=hcfcfb64_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.6=pyhd8ed1ab_0
markupsafe=2.1.2=py310h8d17308_0
mkl=2024.2.2=h66d3029_15
msys2-conda-epoch=20160418=1
multidict=6.0.4=py310h8d17308_0
numpy=1.24.3=py310hd02465a_0
oauthlib=3.2.2=pyhd8ed1ab_1
opencv=4.6.0=py310ha36de5b_5
openssl=1.1.1t=hcfcfb64_0
opt_einsum=3.4.0=pyhd8ed1ab_1
packaging=24.2=pyhd8ed1ab_2
pcre2=10.42=h0ff8eda_1
pip=25.0=py310haa95532_0
platformdirs=4.3.6=pyhd8ed1ab_1
pooch=1.8.2=pyhd8ed1ab_1
protobuf=3.20.3=py310h5588dad_1
pthreads-win32=2.9.1=hfa6e2cd_3
pyasn1=0.6.1=pyhd8ed1ab_2
pyasn1-modules=0.4.1=pyhd8ed1ab_1
pycparser=2.22=pyh29332c3_1
pyjwt=2.10.1=pyhd8ed1ab_0
pyopenssl=23.2.0=pyhd8ed1ab_1
pysocks=1.7.1=pyh09c184e_7
python=3.10.11=h966fe2a_2
python-flatbuffers=25.2.10=pyhbc23db3_0
python_abi=3.10=2_cp310
pyu2f=0.1.5=pyhd8ed1ab_1
qt-main=5.15.2=he8e5bd7_8
qt-webengine=5.15.9=h5bd16bc_7
qtwebkit=5.212=h2bbfb41_5
re2=2022.04.01=h0e60522_0
requests=2.32.3=pyhd8ed1ab_1
requests-oauthlib=2.0.0=pyhd8ed1ab_1
rsa=4.9=pyhd8ed1ab_1
scipy=1.10.1=py310h578b7cb_0
setuptools=75.8.0=py310haa95532_0
six=1.17.0=pyhd8ed1ab_0
snappy=1.1.10=hfb803bf_0
sqlite=3.45.3=h2bbff1b_0
sqlparse=0.5.2=py310haa95532_0
tbb=2021.9.0=h91493d7_0
tensorboard=2.10.1=pyhd8ed1ab_0
tensorboard-data-server=0.6.1=py310h5588dad_4
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.10.0=mkl_py310hd99672f_0
tensorflow-base=2.10.0=mkl_py310h6a7f48e_0
tensorflow-estimator=2.10.0=py310haa95532_0
termcolor=2.5.0=pyhd8ed1ab_1
tk=8.6.14=h0416ee5_0
typing-extensions=4.12.2=py310haa95532_0
typing_extensions=4.12.2=py310haa95532_0
tzdata=2025a=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_1
urllib3=2.3.0=pyhd8ed1ab_0
vc=14.42=haa95532_4
vs2015_runtime=14.42.34433=he0abc0d_4
werkzeug=3.1.3=pyhd8ed1ab_1
wheel=0.45.1=py310haa95532_0
win_inet_pton=1.1.0=pyh7428d3b_8
wrapt=1.15.0=py310h8d17308_0
xz=5.6.4=h4754444_1
yarl=1.9.1=py310h8d17308_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h8cc25b3_1
zstandard=0.19.0=py310h8d17308_0
zstd=1.5.6=h8880b57_0
